+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.2] 2025-07-04
+
+### Changed
+ - testing dataset query options
+ - add missing nloc_0 arg to dataset queries 
+ 
 ## [1.1.1] 2025-06-27
 ### Added
  - new `toshi_hazard_store.query.datasets` module